requirements.txt
django-modelclone